server/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
mysqlclient==2.1.1
django-redis==5.2.0
pycryptodomex==3.17
djangorestframework-simplejwt==5.2.2
celery==5.2.7
django-celery-results==2.5.1
requests==2.30.0
tqdm==4.65.0